+++ src/minedotcpp/common/map.h
@@ -14,20 +14,20 @@ namespace minedotcpp
 		class MINE_API map
 		{
 		public:
-			unsigned short width;
-			unsigned short height;
+			int width;
+			int height;
 			int remaining_mine_count;
 			cell* cells;
 			neighbour_cache_entry* neighbour_cache;
 
-			map(unsigned short width, unsigned short height, int remaining_mine_count = -1);
+			map(int width, int height, int remaining_mine_count = -1);
 			~map();
 
-			cell* cell_get(unsigned short x, unsigned short y);
+			cell* cell_get(int x, int y);
 			cell* cell_get(point pt);
-			cell* cell_try_get(unsigned short x, unsigned short y);
+			cell* cell_try_get(int x, int y);
 			cell* cell_try_get(point pt);
-			bool cell_exists(unsigned short x, unsigned short y);
+			bool cell_exists(int x, int y);
 			bool cell_exists(point pt);
 			neighbour_cache_entry* neighbour_cache_get(int x, int y);
 			neighbour_cache_entry* neighbour_cache_get(point pt);
